backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.350
pydantic>=2.5.2,<3.0.0
mistralai>=0.0.11,<0.0.12
langchain-mistralai==0.0.3
python-multipart==0.0.6
python-docx==1.0.1
pypdf2==3.0.1